src/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
httptools==0.7.1
idna==3.11
NetSDK @ file:///home/zazaka/Dmain/PB/AccessControl/NetSDK_V3.060.0000002.0.R.250928/dist/NetSDK-2.0.0.1-py3-none-linux_x86_64.whl
pydantic==2.12.3
pydantic_core==2.41.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
